seepalaya-new-celery/src/requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
Django==4.2
django-cors-headers==4.3.1
django-jet-reboot==1.3.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf_api_logger==1.1.15
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
elasticsearch==7.17
elasticsearch-dsl==7.4.1
django-modeltranslation==0.19.2